[BE] [UPDATE] update linter rules and remove unnecessary files
