+++ backend/package.json
@@ -17,7 +17,7 @@
     "build": "npx tsc",
     "prestart": "npm run build",
     "lint": "eslint . --ext .ts",
-    "start:app": "ts-node ./src/app.ts",
+    "start:app": "ts-node ./src/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "./node_modules/nodemon/bin/nodemon.js"
   },
