--- backend/src/assets/restaurants.json
@@ -1,214 +1,0 @@
-{
-    "restaurants" : [
-        {
-            "id" : 1,
-            "name" : "mcdonalds",
-            "phoneNumber" : "0123091",
-            "mealType": [
-                {
-                "id": 0,
-                "name": "maindish",
-                "sortId": 1
-            },
-            {
-                "id": 1,
-                "name": "dessert",
-                "sortId": 2
-            },
-            {
-                "id": 2,
-                "name": "drink",
-                "sortId": 3
-            }
-            ],
-            "dishes" : [
-                {
-                    "id": 0,
-                    "name" : "McSunday",
-                    "allergens" : "milk,sugar",
-                    "description" : "bla bla",
-                    "price" : 13,
-                    "category" : {
-                        "extraGroup" : "cheese, olives",
-                        "menuGroup": "dessert",
-                        "foodGroup": "ice,burger"
-                    }
-                },
-                {
-                    "id": 1,
-                    "name" : "Big Mac",
-                    "allergens" : "cheese,salat",
-                    "description" : "the real mc",
-                    "price" : 21,
-                    "category" : {
-                        "extraGroup": "cheese, olives",
-                        "menuGroup": "maindish",
-                        "foodGroup": "burger,salad"
-                    }
-                }
-            ],
-            "location" : {
-                "streetName": "beckerstreet",
-                "streetNumber": "6A",
-                "postalCode": "12570",
-                "country" : "germany"
-            },
-            "extras": [{
-                "id": 0,
-                "name" : "cheese",
-                "allergens" : "milk",
-                "description" : "bla bla",
-                "price" : 1,
-                "category" : {
-                    "extraGroup": "",
-                    "menuGroup": "",
-                    "foodGroup": ""
-                }
-            }]
-        },
-        {
-            "id" : 2,
-            "name" : "burger king",
-            "phoneNumber": "01232191",
-            "mealType": [
-                {
-                    "id": 0,
-                    "name": "maindish",
-                    "sortId": 2
-                },
-                {
-                    "id": 1,
-                    "name": "dessert",
-                    "sortId": 1
-                },
-                {
-                    "id": 2,
-                    "name": "drink",
-                    "sortId": 3
-                }
-            ],
-            "dishes" : [
-                {
-                    "id": 0,
-                    "name" : "Big Sunday ice",
-                    "allergens" : "milk,sugar",
-                    "description" : "bla bla",
-                    "price" : 12,
-                    "category" : {
-                        "extraGroup": "cheese, olives",
-                        "menuGroup": "dessert",
-                        "foodGroup": "burger, ice"
-                    }
-                },
-                {
-                    "id": 1,
-                    "name" : "Big Big Mac",
-                    "allergens" : "cheese,salat",
-                    "description" : "the real big mc",
-                    "price" : 5,
-                    "category" : {
-                        "extraGroup": "cheese, olives",
-                        "menuGroup": "dessert",
-                        "foodGroup": "burger,salad"
-                    }
-                }
-            ],
-            "location" : {
-                "streetName": "Notbeckerstreet",
-                "streetNumber": "10",
-                "postalCode": "12123",
-                "country" : "germany"
-            },
-            "extras": [{
-                "id": 0,
-                "name" : "cheese",
-                "allergens" : "milk",
-                "description" : "bla bla",
-                "price" : 1,
-                "category" : {
-                    "extraGroup": "",
-                    "menuGroup": "",
-                    "foodGroup": ""
-                }
-            }]
-        },
-        {
-            "id" : 3,
-            "name" : "burgerme",
-            "phoneNumber": "0123091123",
-            "mealType": [
-                {
-                    "id": 0,
-                    "name": "maindish",
-                    "sortId": 3
-                },
-                {
-                    "id": 1,
-                    "name": "dessert",
-                    "sortId": 1
-                },
-                {
-                    "id": 2,
-                    "name": "appetizer",
-                    "sortId": 2
-                }
-            ],
-            "dishes" : [
-                {
-                    "id": 0,
-                    "name" : "BurgerMega Sunday Ice",
-                    "allergens" : "sugar",
-                    "description" : "bla bla",
-                    "price" : 12,
-                    "category" : {
-                        "extraGroup": "cheese, olives",
-                        "menuGroup": "dessert",
-                        "foodGroup": "burger"
-                    }
-                },
-                {
-                    "id": 1,
-                    "name" : "Monster Salad",
-                    "allergens" : "sugar",
-                    "description" : "bla bla",
-                    "price" : 10,
-                    "category" : {
-                        "extraGroup": "cheese, olives",
-                        "menuGroup": "appetizer",
-                        "foodGroup": "burger"
-                    }
-                }
-            ],
-            "location" : {
-                "streetName": "beckeraaastreet",
-                "streetNumber": "96A",
-                "postalCode": "12590",
-                "country" : "germany"
-            },
-            "extras": [{
-                "id": 0,
-                "name" : "cheese",
-                "allergens" : "milk",
-                "description" : "bla bla",
-                "price" : 1,
-                "category" : {
-                    "extraGroup": "",
-                    "menuGroup": "",
-                    "foodGroup": ""}
-            },
-                {
-                "id": 1,
-                "name" : "morzerella",
-                "allergens" : "milk",
-                    "description" : "bla bla",
-                    "price" : 2,
-                    "category" : {
-                        "extraGroup": "",
-                        "menuGroup": "",
-                        "foodGroup": ""
-                    }
-                }
-            ]
-        }
-    ]
-}
