+++ backend/src/index.ts
@@ -1,12 +1,38 @@
-import * as express from 'express';
-import filter from './routes/filter';
-const router = express.Router();
+import createError from 'http-errors';
+import express from 'express';
+import cookieParser from 'cookie-parser';
+import logger from 'morgan';
+import path = require('path');
 
-/* GET home page. */
-router.get('/', function (req, res) {
-  res.send('Hello World!!!!');
-});
+async function main() {
+  const app = express();
+  const port = 8081;
 
-router.get('/api/filter', filter);
+  app.use(logger('dev'));
+  app.use(express.json());
+  app.use(express.urlencoded({ extended: false }));
+  app.use(cookieParser());
+  app.use(express.static(path.join(__dirname, 'public')));
 
-export default router;
+  // catch 404 and forward to error handler
+  app.use(function(req: any, res: any, next: any) { /* eslint-disable-line */
+    next(createError(404));
+  });
+
+  // error handler
+  app.use(function(err: any, req: any, res: any) { /* eslint-disable-line */
+    // set locals, only providing error in development
+    res.locals.message = err.message;
+    res.locals.error = req.app.get('env') === 'development' ? err : {};
+
+    // render the error page
+    res.status(err.status || 500);
+    res.render('error');
+  });
+}
+
+try {
+  main();
+} catch (err: unknown) {
+  console.error(err);
+}
